Move loader detection into `load_context` and `contrib` modules
