+++ crates/oak_db/src/lib.rs
@@ -6,14 +6,13 @@ mod file;
 mod file_diagnostics;
 mod file_exports;
 mod file_imports;
-mod file_load_context;
 mod file_resolve;
 mod file_revision;
-mod file_shiny;
 mod file_source_site;
 mod identifier;
 mod imports;
 mod inputs;
+mod load_context;
 mod name;
 mod package;
 mod package_resolve;
