+++ crates/oak_db/src/load_context/contrib.rs
@@ -1,0 +1,4 @@
+//! Package-defined file-loading conventions.
+
+pub(crate) mod shiny;
+pub(crate) mod testthat;
