+++ crates/oak_db/src/tests/contrib.rs
@@ -1,0 +1,2 @@
+mod shiny;
+mod testthat;
